chore: Proxy health-check endpoint improvements (#3482)
* chore: proxy healtcheck improvements, work in progress

Refs: XRDDEV-3078

* chore: maintenance mode filter for healthcheck

Refs: XRDDEV-3078

* chore: change proxy healthcheck port 4099->5588

Refs: XRDDEV-3078

* chore: make k8 readinessProbe fields configurable with defaults

Refs: XRDDEV-3078

* chore: trailing '/' fix in health endpoint

Refs: XRDDEV-3078

* chore: drop HealthCheckPortImpl

Refs: XRDDEV-3078

* chore: use custom HeapMemory check for proxy

Refs: XRDDEV-3078

* chore: ss system tests fix

Refs: XRDDEV-3078

* chore: ss system tests fix

Refs: XRDDEV-3078

* chore: add a way to create a new property based on migrated property value

Refs: XRDDEV-3078

* chore: caching healthcheck fix, cleanup

Refs: XRDDEV-3078

* chore: use injected clock in CachingHealthCheckTest

Refs: XRDDEV-3078

* chore: update configurable-properties.yaml

Refs: XRDDEV-3078

* chore: fix @GetMapping annotation in FeignHealthcheckApi

Refs: XRDDEV-3078

+++ src/lib/health-check-core/src/main/java/org/niis/xroad/common/healthcheck/HeapMemoryLivenessCheck.java
@@ -68,7 +68,7 @@ public class HeapMemoryLivenessCheck implements HealthCheck {
         return HealthCheckResponse.named(NAME)
                 .up()
                 .withData("used_percent", usedPercent)
-                .withData("threshold_percent", (long) THRESHOLD_PERCENT)
+                .withData("threshold_percent", THRESHOLD_PERCENT)
                 .build();
     }
